Exportação de horário para PDF
Com alterações, ao longo do tempo a classe que tinha a exportação do horário foi mudada e mudado o nome, o que provocou com que a exportação de horários deixasse de funcionar. Eliminiação de classe duplicada
